Note unnecessary configuration for Heroku
I recently attempted an upgrade of Postgres on Heroku, and had it fail because of this schema extension. The `pg:upgrade` command was in no hurry to tell me why so we incurred some longer downtime than was necessary. Then when re-adding this schema was reminded of a few steps here that did not seem to apply to database instances there that I figured were worth calling out for others deploying there.