Simplify the presentation
Present a explanation box recommending to use CONCURRENTLY with DROP
INDEX. In Active Record, this requires two things:
disable_ddl_transaction! once at the top of the migration file, and
"algorithm: :concurrently" option added to every remove_index line

Since this is a sensible default when PostgreSQL 14 or greater is in
use, since it can avoid downtime from an index drop, make it the
default.
Remove the ability to disable it since it seems unnecessary in
retrospect.

Only show the explanation box and only add the option to
remove_index when version 14 or greater is in use